// src/app/entities/vehiculo-interface.ts
export interface VehiculoInterface{

    id?: number;
    placa?: string;
    descripcion?: string;
    chasis?: string;
    marca?: string;
    codigo_patrimonial?: string;
    peso_bruto?: number;
    peso_neto?: number;
    asientos?: number;
    anio_fabricacion?: number;
    modelo?: string;
    color?: string;
    cantidad_cilindro?: number;
    id_tipo_combustible?: number;
    id_rendimiento?: number;
    numero_orden?: string;
    fecha_compra?: string;
    proveedor?: string;
    anio_compra?: number;
    valor_compra?: number;
    usuario_registro?: number;
    nombre_usuario?: string;
    fecha_registro?: string;
    estado?: number;
    nombre_estado?: string;
    id_tipo_vehiculo?: number;
    nombre_vehiculo?: string;
    kilometraje?:number;

    id_organo?: number;
    responsable?: string;
} 